requirements.txt
camelot-py==0.9.0
cffi==1.15.0
chardet==4.0.0
click==8.1.3
cryptography==37.0.1
et-xmlfile==1.1.0
numpy==1.22.3
opencv-python==4.5.5.64
openpyxl==3.0.9
pandas==1.4.2
pdfminer.six==20220319
pycparser==2.21
PyPDF2==1.27.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
six==1.16.0
